pithos: Handle XF_ZERO_FLAG on archipelago mapper
Handle XF_ZERO_FLAG on archipelago mapper. Instead of explicitly creating and/or
refering to the deduplicated zero object, mark the zero object with a dedicated
flag and handle any reference to it accordingly. For example there is no need to
actually access the zero object on disk, when we can simply pad the buffer with
zeroes. Moreover when creating new files, mark any reference to the zero object
with the dedicated flag.